fix: add StrainRef, to bridge between the db tables and pydantic
`StrainRef` serves to keep a reference to the ID the strain has in BRENDA,
since that ID will also be the the doc_id the strain will receive in our
document database.

`StrainRef` is declared as a NamedTuple to avoid declaring a new Pydantic
model.

This whole refactoring was necessary to fix the problems with SQLModel not
working well with a primary key combined with a field alias (necessary to map
the id field to BRENDA's protein_organism_strain_id).